LPD-24485 Inline JSONWebServiceTrackerCustomizer into JSONWebServiceActionsManagerImpl
